emotional-chronicler/requirements.txt
# The Emotional Chronicler — Python dependencies
#
# Core framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Google ADK (Agent Development Kit — orchestration, tool calling, session management)
google-adk>=0.4.0

# Google GenAI SDK (Imagen 4 image generation, Lyria 2 music generation)
google-genai>=1.0.0

# Authentication & environment
google-auth>=2.35.0
python-dotenv>=1.0.0

# Firebase Authentication
firebase-admin>=6.0.0

# Firestore (session memory)
google-cloud-firestore>=2.19.0
